+++ Libraries/LibWeb/Encoding/TextDecoderCommon.idl
@@ -1,0 +1,6 @@
+// https://encoding.spec.whatwg.org/#textdecodercommon
+interface mixin TextDecoderCommon {
+    readonly attribute DOMString encoding;
+    readonly attribute boolean fatal;
+    readonly attribute boolean ignoreBOM;
+};
