+++ Libraries/LibWeb/idl_files.cmake
@@ -578,6 +578,7 @@ libweb_support_idl(DOM/EventHandler)
 libweb_support_idl(DOM/EventListener)
 libweb_support_idl(DOM/ParentNode)
 libweb_support_idl(DOM/Slottable)
+libweb_support_idl(Encoding/TextDecoderCommon)
 libweb_support_idl(Encoding/TextEncoderCommon)
 libweb_support_idl(Fetch/Body)
 libweb_support_idl(Fetch/BodyInit)
